fix: allow unused vars from destructuring (#22)
Adds `ignoreRestSiblings: true` to the `no-unused-vars` rule options. This is important when you want to destructure an object and remove some keys from it, e.g. `const { opt1, ...rest } = options;` so you can pass down just the `rest` and omit properties that shouldn't be there. We use this quite extensively.

https://eslint.org/docs/latest/rules/no-unused-vars#ignorerestsiblings