feat: improve import sorting, add `node:` prefix (#27)
This PR adds a new logic for import sorting as import/order doesn't work properly and the behavior is considered correct.
We also add a rule to enforce node: prefix for modules so we can have nice sorting. (I had to release eslint-plugin-import on my own + fix a bug there, that's why the weird package 😅 )