chore(ci): Rewrite publish workflow (#34)
This rewrites the publishing workflow to:
- use OIDC to publish to NPM
- bump the package version based on the workflow input
- automatically generate the changelog
- commit the version bump and changelog
- generate a GitHub release

One problem is that since the OIDC publishing to NPM allows only one
publishing workflow, we lose the option to publish beta versions, but
since this package is mostly stable, it's not really necessary.

I've tested everything but the actual NPM publishing, I don't want to
generate unnecessary package versions.